# Backend Files/SplitGoFilesVersion/docs/swagger.yaml
basePath: /
definitions:
  main.Coordinate:
    properties:
      latitude:
        type: number
      longitude:
        type: number
    type: object
info:
  contact: {}
  description: APIs for frontend use.
  title: Backend APIs
  version: "1.0"
paths:
  /location:
    get:
      description: Retrieve the latest stored location
      produces:
      - text/plain
      responses:
        "200":
          description: Latest location details
          schema:
            type: string
      summary: Get latest location
      tags:
      - Location API
    put:
      consumes:
      - application/json
      description: Update location with latitude and longitude
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: string
      - description: Longitude
        in: query
        name: lng
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latitude and Longitude updated successfully
          schema:
            type: string
      summary: Update location
      tags:
      - Location API
  /maps/coordstosn:
    get:
      description: Convert latitude and longitude to street name
      parameters:
      - description: Latitude
        in: query
        name: lat
        required: true
        type: string
      - description: Longitude
        in: query
        name: long
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Street name
          schema:
            type: string
      summary: Convert coordinates to street name
      tags:
      - Maps API
  /maps/sntocoords:
    get:
      consumes:
      - application/json
      description: Convert street name to latitude and longitude
      parameters:
      - description: Street name
        in: query
        name: streetName
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Latitude and Longitude
          schema:
            items:
              type: number
            type: array
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Convert street name to coordinates
      tags:
      - Maps API
  /recommendations:
    get:
      consumes:
      - application/json
      description: Retrieves a list of recommended locations based on user's current
        position and other parameters.
      parameters:
      - description: Latitude of the user's current position
        in: query
        name: lat
        required: true
        type: number
      - description: Longitude of the user's current position
        in: query
        name: lon
        required: true
        type: number
      - description: Number of random coordinates to generate around the user's position
        in: query
        name: numOfRandomCoordinates
        type: integer
      - description: Maximum number of recommended locations to return (at least 1
          will be returned)
        in: query
        name: maxNumOfResults
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: A list of recommended locations with latitude, longitude, and
            estimated travel duration in seconds
          schema:
            items:
              type: array
            type: array
        "400":
          description: Bad Request
          schema:
            type: string
        "500":
          description: Internal Server Error
          schema:
            type: string
      summary: Get Recommended Locations
      tags:
      - Recommendation API
  /taxi:
    get:
      description: Retrieve taxi coordinates from an external API and return only
        those within 5km of the specified coordinate as JSON.
      parameters:
      - description: Latitude of the reference point
        in: query
        name: latitude
        required: true
        type: number
      - description: Longitude of the reference point
        in: query
        name: longitude
        required: true
        type: number
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/main.Coordinate'
            type: array
      summary: Get taxi coordinates within 5km of a given coordinate
      tags:
      - Taxi API
swagger: "2.0"